Compute the expected number of events using lumi masks
In case the workflow provides a `LumiList`, use it to compute the
expected number of events in case they are not available in the workflow
dictionary yet.

This is aimed towards computing properly the `expected` events before
the value is defined when the workflow reaches the `assigned`
status as it is required for monitoring and operations purposes.

Signed-off-by: Geovanny González-Rodríguez <geovanny.gonzalez@cern.ch>